Discord plugin - First pass
A work in progress on a Discord bot plugin. Right now, it's only capable of pushing notifications to a channel (no commands). It doesn't have any settings either.

Relies on the excellent and stable discord.js project

What it does if you enable it as is:

- Notify you when a trade is completed
- Notify you on performance report

Screenshot of current implementation's result: https://i.imgur.com/0fy2NgE.png